requirements.txt
fastapi
uvicorn
streamlit
sentence-transformers
faiss-cpu
langchain
pydantic
python-multipart
PyPDF2
transformers
torch
numpy
pandas
requests
python-dotenv
PyYAML
python-docx
openai